Adding xib connections to screens in abstract super classes.

+++ jos/src/main/java/jos/api/graphicsimaging/CGPoint.java
@@ -4,10 +4,10 @@ import com.google.j2objc.annotations.Register;
 
 @Register(isWrapper = true)
 public class CGPoint {
-    public CGFloat x;
-    public CGFloat y;
+    public float x;
+    public float y;
 
-    public CGPoint(final CGFloat x, final CGFloat y) {
+    public CGPoint(final float x, final float y) {
         this.x = x;
         this.y = y;
     }
